requirements.txt
torch>=2.0.0
tiktoken>=0.5.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.28.0
flask>=2.0.0
PyYAML>=6.0
streamlit>=1.30.0
